dockershim: clean up unused security context code
Also remove references to kubernetes api objects
