+++ pkg/kubelet/dockershim/BUILD
@@ -27,6 +27,7 @@ go_library(
         "helpers_linux.go",
         "naming.go",
         "security_context.go",
+        "selinux_util.go",
     ],
     tags = ["automanaged"],
     deps = [
@@ -39,7 +40,6 @@ go_library(
         "//pkg/kubelet/dockershim/cm:go_default_library",
         "//pkg/kubelet/dockershim/errors:go_default_library",
         "//pkg/kubelet/dockershim/libdocker:go_default_library",
-        "//pkg/kubelet/dockershim/securitycontext:go_default_library",
         "//pkg/kubelet/leaky:go_default_library",
         "//pkg/kubelet/network:go_default_library",
         "//pkg/kubelet/network/cni:go_default_library",
@@ -86,6 +86,7 @@ go_test(
         "helpers_test.go",
         "naming_test.go",
         "security_context_test.go",
+        "selinux_util_test.go",
     ],
     data = [
         "fixtures/seccomp/sub/subtest",
@@ -99,7 +100,6 @@ go_test(
         "//pkg/kubelet/container/testing:go_default_library",
         "//pkg/kubelet/dockershim/errors:go_default_library",
         "//pkg/kubelet/dockershim/libdocker:go_default_library",
-        "//pkg/kubelet/dockershim/securitycontext:go_default_library",
         "//pkg/kubelet/dockershim/testing:go_default_library",
         "//pkg/kubelet/network:go_default_library",
         "//pkg/kubelet/network/testing:go_default_library",
@@ -134,7 +134,6 @@ filegroup(
         "//pkg/kubelet/dockershim/errors:all-srcs",
         "//pkg/kubelet/dockershim/libdocker:all-srcs",
         "//pkg/kubelet/dockershim/remote:all-srcs",
-        "//pkg/kubelet/dockershim/securitycontext:all-srcs",
         "//pkg/kubelet/dockershim/testing:all-srcs",
     ],
     tags = ["automanaged"],
